chore: move the stateful transaction validator config into config file
