+++ crates/gateway/tests/routing_test.rs
@@ -9,10 +9,11 @@ use blockifier::blockifier::block::BlockInfo;
 use blockifier::test_utils::dict_state_reader::DictStateReader;
 use pretty_assertions::assert_str_eq;
 use rstest::rstest;
-use starknet_gateway::config::{GatewayNetworkConfig, StatelessTransactionValidatorConfig};
+use starknet_gateway::config::{
+    GatewayNetworkConfig, StatefulTransactionValidatorConfig, StatelessTransactionValidatorConfig,
+};
 use starknet_gateway::gateway::Gateway;
 use starknet_gateway::state_reader_test_utils::{TestStateReader, TestStateReaderFactory};
-use starknet_gateway::stateful_transaction_validator::StatefulTransactionValidatorConfig;
 use starknet_mempool_types::mempool_types::{
     GatewayNetworkComponent, GatewayToMempoolMessage, MempoolToGatewayMessage,
 };
